requirements.txt
onnxruntime
flask
opencv-python
pillow